requirements.txt
pandas
numpy
pdfminer.six
pytesseract
Wand
Pillow
PyMuPDF
polyglot
pyicu
pycld2
morfessor
python-docx
python-pptx